more renamings and shufflinf stuff around
