+++ webdriver-examples/README.md
@@ -3,7 +3,7 @@
 - [webdriver-examples](#webdriver-examples)
   - [About These Examples](#about-these-examples)
   - [Core Modules](#core-modules)
-    - [IORunner](#iorunner)
+    - [HttpRunner](#HttpRunner)
     - [IOAPI](#ioapi)
     - [HttpE2EDemoTest](#httpe2edemotest)
   - [Running Examples (VSCode Dev-Container)](#running-examples-vscode-dev-container)
@@ -34,9 +34,9 @@ The examples cover all [W3C WebDriver endpoints](https://www.w3.org/TR/webdriver
 
 ## Core Modules
 
-### IORunner
+### HttpRunner
 
-[IORunner](./driver-demo-e2e/IORunner.hs) exports a single `run` function that accepts a [webdriver-precore HttpSpec](https://hackage-content.haskell.org/package/webdriver-precore-0.0.0.1/package/docs/WebDriverPreCore.html#g:14) and performs HTTP requests to an active WebDriver instance.
+[HttpRunner](./driver-demo-e2e/HttpRunner.hs) exports a single `run` function that accepts a [webdriver-precore HttpSpec](https://hackage-content.haskell.org/package/webdriver-precore-0.0.0.1/package/docs/WebDriverPreCore.html#g:14) and performs HTTP requests to an active WebDriver instance.
 
 ### IOAPI
