+++ webdriver-examples/webdriver-examples.cabal
@@ -21,7 +21,7 @@ synopsis:       A test suite that demonstrates how to use the webdriver-precore 
 
 description:
   The E2E tests in HttpE2EDemoTest module demonstrate how the webdriver-precore might be used in a library that provides a basic runner.
-  The runner is implemented in IORunner module. See the @<https://github.com/pyrethrum/webdriver/webdriver-examples#readme/ the README>@ for more details.
+  The runner is implemented in HttpRunner module. See the @<https://github.com/pyrethrum/webdriver/webdriver-examples#readme/ the README>@ for more details.
 
 
 common commonExtensions
