+++ webdriver-precore/src/WebDriverPreCore/Internal/AesonUtils.hs
@@ -1,0 +1,92 @@
+module WebDriverPreCore.Internal.AesonUtils
+  ( aesonTypeError,
+    aesonTypeErrorMessage,
+    asText,
+    bodyText',
+    lookup,
+    lookupTxt,
+    opt,
+    jsonPrettyString,
+    jsonToText,
+    prettyPrintJson,
+    parseJson
+  )
+where
+
+-- \| Utility functions for working with Aeson (JSON) values.
+import Control.Exception (Exception (displayException), SomeException, try)
+import Data.Aeson
+  ( Key,
+    KeyValue ((.=)),
+    Result (..),
+    ToJSON (),
+    Value (..),
+    eitherDecodeStrict,
+  )
+import Data.Aeson.Encode.Pretty (encodePretty)
+import Data.Aeson.KeyMap qualified as AKM
+import Data.ByteString.Lazy qualified as LBS
+import Data.Either (Either, either)
+import Data.Function (($), (&), (.))
+import Data.Functor (Functor, (<$>))
+import Data.String (String)
+import Data.Text (Text, unpack)
+import Data.Text.Encoding (decodeUtf8, encodeUtf8)
+import Data.Text.IO qualified as T
+import GHC.Base (IO)
+import GHC.Show (Show (..))
+import System.IO (print)
+import Prelude
+  ( maybe,
+    pure,
+    (<>),
+    (>>=),
+  )
+
+-- Aeson stuff
+-- TODO move to separte library
+
+asText :: Value -> Result Text
+asText = \case
+  String t -> Success t
+  v -> aesonTypeError "Text" v
+
+opt :: (Functor f, KeyValue e b, ToJSON a) => Key -> f a -> f b
+opt lbl mb = (lbl .=) <$> mb
+
+-- https://blog.ssanj.net/posts/2019-09-24-pretty-printing-json-in-haskell.html
+lsbToText :: LBS.ByteString -> Text
+lsbToText = decodeUtf8 . LBS.toStrict
+
+jsonToText :: Value -> Text
+jsonToText = lsbToText . encodePretty
+
+prettyPrintJson :: Value -> IO ()
+prettyPrintJson v = do
+  e <- (try @SomeException @_) $ T.putStrLn (jsonToText v)
+  either (print . displayException) print e
+
+parseJson :: Text -> Either String Value
+parseJson input =
+  eitherDecodeStrict (encodeUtf8 input)
+
+bodyText' :: Result Value -> Key -> Result Text
+bodyText' v k = v >>= lookupTxt k
+
+aesonTypeErrorMessage :: Text -> Value -> Text
+aesonTypeErrorMessage t v = "Expected JSON Value to be of type: " <> t <> "\nbut got:\n" <> jsonToText v
+
+aesonTypeError :: Text -> Value -> Result a
+aesonTypeError t v = Error . unpack $ aesonTypeErrorMessage t v
+
+jsonPrettyString :: Value -> String
+jsonPrettyString = unpack . jsonToText
+
+lookupTxt :: Key -> Value -> Result Text
+lookupTxt k v = lookup k v >>= asText
+
+lookup :: Key -> Value -> Result Value
+lookup k v =
+  v & \case
+    Object o -> AKM.lookup k o & maybe (Error ("the key: " <> show k <> "does not exist in the object:\n" <> jsonPrettyString v)) pure
+    _ -> aesonTypeError "Object" v
