+++ webdriver-precore/src/WebDriverPreCore/Internal/Utils.hs
@@ -1,69 +1,52 @@
 {-# OPTIONS_HADDOCK hide #-}
 
 module WebDriverPreCore.Internal.Utils
-  ( opt,
-    txt,
+  ( txt,
     enumerate,
-    -- Aeson utils
-    jsonToText,
-    lsbToText,
-    prettyPrintJson,
-    parseJson
+    -- shared path elements
+    UrlPath (..),
+    newSessionUrl,
+    session,
+    bodyValue,
+    bodyText
   )
 where
 
-import Data.Aeson
-  ( Key,
-    KeyValue ((.=)),
-    ToJSON (),
-    Value (..), eitherDecodeStrict,
-  )
-import Data.Function ((.), ($))
-import Data.Functor (Functor, (<$>))
-import Data.String (String)
-import Data.Text ( Text, pack)
-import Data.Text.Encoding (encodeUtf8, decodeUtf8)
+import Data.Function ((.))
+import Data.Text (Text, pack)
+import GHC.Enum (Bounded (..), Enum)
 import GHC.Show (Show (..))
-import Data.Aeson.Encode.Pretty (encodePretty)
-import Data.ByteString.Lazy qualified as LBS
-import Data.Text.IO qualified as T
-import Control.Exception (try, Exception (displayException), SomeException)
-import GHC.Base (IO)
-import Data.Either (Either, either)
-import System.IO (print)
-import GHC.Enum (Enum, Bounded (..))
-
+import Prelude (Eq, Ord, Semigroup)
+import WebDriverPreCore.Http.HttpResponse (HttpResponse(..))
+import Data.Aeson (Key, Result, Value)
+import WebDriverPreCore.Internal.AesonUtils (lookup, bodyText')
 
 {-
   this module is used between the library and testing modules
-  it will be removed in a later relaease
+  it will be removed in a later release
 -}
+
+-- general utils
 
 txt :: (Show a) => a -> Text
 txt = pack . show
 
-
 enumerate :: (Enum a, Bounded a) => [a]
 enumerate = [minBound ..]
 
--- Aeson stuff
--- TODO move to separte library
+-- shared path elements
+newtype UrlPath = MkUrlPath {segments :: [Text]}
+  deriving newtype (Show, Eq, Ord, Semigroup)
 
-opt :: (Functor f, KeyValue e b, ToJSON a) => Key -> f a -> f b
-opt lbl mb = (lbl .=) <$> mb
+newSessionUrl :: UrlPath
+newSessionUrl = MkUrlPath [session]
 
--- https://blog.ssanj.net/posts/2019-09-24-pretty-printing-json-in-haskell.html
-lsbToText :: LBS.ByteString -> Text
-lsbToText = decodeUtf8 . LBS.toStrict
+session :: Text
+session = "session"
 
-jsonToText :: Value -> Text
-jsonToText = lsbToText . encodePretty
+bodyValue :: HttpResponse -> Result Value
+bodyValue r = lookup "value" r.body
 
-prettyPrintJson :: Value -> IO ()
-prettyPrintJson v = do
-  e <- (try @SomeException @_) $ T.putStrLn (jsonToText v)
-  either (print . displayException) print e
+bodyText :: HttpResponse -> Key -> Result Text
+bodyText r = bodyText' (bodyValue r)
 
-parseJson :: Text -> Either String Value
-parseJson input =
-  eitherDecodeStrict (encodeUtf8 input)
